# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Elkanawy OS Smart Home API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface User {
  id: number;
  googleId?: string;
  email: string;
  name: string;
  avatarUrl?: string | null;
  createdAt: string;
}

export interface GoogleSignInRequest {
  credential: string;
  clientId?: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Room {
  id: number;
  name: string;
  icon: string;
  userId: number;
  deviceCount: number;
  activeCount: number;
  createdAt: string;
}

export interface CreateRoomRequest {
  name: string;
  icon: string;
}

export type DeviceType = (typeof DeviceType)[keyof typeof DeviceType];

export const DeviceType = {
  light: "light",
  fan: "fan",
  ac: "ac",
  thermostat: "thermostat",
  lock: "lock",
  camera: "camera",
  speaker: "speaker",
  tv: "tv",
} as const;

export interface Device {
  id: number;
  name: string;
  type: DeviceType;
  isOn: boolean;
  roomId: number;
  userId: number;
  /** Numeric value (brightness, temperature, fan speed, etc.) */
  value?: number | null;
  /** Unit for value (%, °C, etc.) */
  unit?: string | null;
  powerUsage?: number | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateDeviceRequestType =
  (typeof CreateDeviceRequestType)[keyof typeof CreateDeviceRequestType];

export const CreateDeviceRequestType = {
  light: "light",
  fan: "fan",
  ac: "ac",
  thermostat: "thermostat",
  lock: "lock",
  camera: "camera",
  speaker: "speaker",
  tv: "tv",
} as const;

export interface CreateDeviceRequest {
  name: string;
  type: CreateDeviceRequestType;
  roomId: number;
  value?: number | null;
  unit?: string | null;
}

export interface UpdateDeviceRequest {
  name?: string;
  isOn?: boolean;
  value?: number | null;
  unit?: string | null;
}

export type AutomationAction =
  (typeof AutomationAction)[keyof typeof AutomationAction];

export const AutomationAction = {
  turn_on: "turn_on",
  turn_off: "turn_off",
  set_value: "set_value",
} as const;

export type AutomationTriggerType =
  (typeof AutomationTriggerType)[keyof typeof AutomationTriggerType];

export const AutomationTriggerType = {
  schedule: "schedule",
  sunset: "sunset",
  sunrise: "sunrise",
  manual: "manual",
} as const;

export interface Automation {
  id: number;
  name: string;
  deviceId: number;
  action: AutomationAction;
  triggerType: AutomationTriggerType;
  /** Time in HH:MM format for schedule triggers */
  triggerTime?: string | null;
  /** Comma-separated days (mon,tue,wed...) */
  triggerDays?: string | null;
  value?: number | null;
  isEnabled: boolean;
  userId: number;
  createdAt: string;
}

export type CreateAutomationRequestAction =
  (typeof CreateAutomationRequestAction)[keyof typeof CreateAutomationRequestAction];

export const CreateAutomationRequestAction = {
  turn_on: "turn_on",
  turn_off: "turn_off",
  set_value: "set_value",
} as const;

export type CreateAutomationRequestTriggerType =
  (typeof CreateAutomationRequestTriggerType)[keyof typeof CreateAutomationRequestTriggerType];

export const CreateAutomationRequestTriggerType = {
  schedule: "schedule",
  sunset: "sunset",
  sunrise: "sunrise",
  manual: "manual",
} as const;

export interface CreateAutomationRequest {
  name: string;
  deviceId: number;
  action: CreateAutomationRequestAction;
  triggerType: CreateAutomationRequestTriggerType;
  triggerTime?: string | null;
  triggerDays?: string | null;
  value?: number | null;
}

export interface UpdateAutomationRequest {
  name?: string;
  isEnabled?: boolean;
  triggerTime?: string | null;
  triggerDays?: string | null;
  value?: number | null;
}

export type GetDevicesParams = {
  roomId?: number;
};