# lib/api-zod/src/generated/types/automation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Elkanawy OS Smart Home API
 * OpenAPI spec version: 0.1.0
 */
import type { AutomationAction } from "./automationAction";
import type { AutomationTriggerType } from "./automationTriggerType";

export interface Automation {
  id: number;
  name: string;
  deviceId: number;
  action: AutomationAction;
  triggerType: AutomationTriggerType;
  /** Time in HH:MM format for schedule triggers */
  triggerTime?: string | null;
  /** Comma-separated days (mon,tue,wed...) */
  triggerDays?: string | null;
  value?: number | null;
  isEnabled: boolean;
  userId: number;
  createdAt: Date;
}