# lib/api-zod/src/generated/types/createAutomationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Elkanawy OS Smart Home API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAutomationRequestAction } from "./createAutomationRequestAction";
import type { CreateAutomationRequestTriggerType } from "./createAutomationRequestTriggerType";

export interface CreateAutomationRequest {
  name: string;
  deviceId: number;
  action: CreateAutomationRequestAction;
  triggerType: CreateAutomationRequestTriggerType;
  triggerTime?: string | null;
  triggerDays?: string | null;
  value?: number | null;
}