# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Elkanawy OS Smart Home API
 * OpenAPI spec version: 0.1.0
 */

export interface Room {
  id: number;
  name: string;
  icon: string;
  userId: number;
  deviceCount: number;
  activeCount: number;
  createdAt: Date;
}