# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Elkanawy OS Smart Home API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  googleId?: string;
  email: string;
  name: string;
  avatarUrl?: string | null;
  createdAt: Date;
}